Implemented CRUD operations except deleting on modal windows.
Implemented CRUD operations except deleting on modal windows. The read
part is made through a modal and the create and update part is made
using another modal. The routing is made using states, ui-router.
Extracted the product form in a separate file, modified the controllers
logic in order to focus on the products list separately from the product
form logic.